backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apscheduler==3.10.4",
    "croniter==6.0.0",
    "cryptography==43.0.0",
    "fastapi==0.112.0",
    "httpx==0.27.0",
    "plexapi==4.15.11",
    "pydantic-settings==2.4.0",
    "pytest==8.3.2",
    "python-levenshtein>=0.27.1",
    "python-multipart==0.0.9",
    "tenacity==8.3.0",
    "thefuzz[speedup]==0.22.1",
    "uvicorn[standard]==0.30.1",
    "alembic==1.13.2",
    "sqlalchemy==2.0.31",
    "mutagen==1.47.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
]
